Added more tests, fixed imports
